iOS, mkspec: remove faulty SDK check
We used to check if the SDK used by a project is less than 8.0 and error
out if so. The intention of the test was to avoid a situation where a
project is built with an SDK that is older than the one used to build Qt,
but this was obviously bogus, as Qt could have been built with a newer SDK
than the oldest supported one. Also, 8.0 has been outdated for quite a
while.

On top of that, the check failed now that the major iOS version has two
digits.

So let's remove the check for now, until we can handle this in a better
way.

See QTBUG-37592.

Change-Id: I6106b9521b5d47d9906d4db30c2ffa21794bc307
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>