ios_destinations.sh: ignore devices that are marked as placeholders
When xcodebuild in Xcode 8 beta dumps out the available
destinations, it prints an extra section called 'Ineligible
destinations for the "tst_someTest" scheme'.

Those destinations doesn't contain valid ID-s for the script to
use, which will result in "make check" failing.

This patch will filter out devices that are marked as placeholders.

Change-Id: I88a25b7307e21b76c6f7764a82f67627aae8f02f
Reviewed-by: Jake Petroules <jake.petroules@qt.io>