Fix QGtk3ColorDialogHelper::currentColorChanged() emission
Even though GtkColorChooser provides a "color-activated" signal, it
does not get emitted whenever the selection in the dialog changes, but
only when a color is activated in a sense that the dialog is accepted.
Switch to the "notify::rgba" signal to get proper live notifications
of the selection changes.

Task-number: QTBUG-55298
Change-Id: I39ab798764c8b42a998d75c56605948eaf154dcd
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Dmitry Shachnev <mitya57@gmail.com>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>