Revert "Cocoa: make dialogs emit the "selected" signals"
This reverts commit dfa8854cf7e00705e0122cf7022ff1ea4f8e5a74 and
removes all "selected" signal emissions from the Cocoa platform
dialogs.

Even though it fixed the new QML dialogs that were relying on the
"selected" signals, it lead to duplicate signals with QColorDialog,
QFileDialog, and QFontDialog. We'll fix the new QML dialogs to not
rely on the selected signals, but handle it on accept the same way
than QtWidgets and QtQuick Dialogs do, so there is no need to repeat
the signals in all platform plugins.

Task-number: QTBUG-55299
Change-Id: I35e08cee92a4454544497b027ed10abad6c26673
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>